live-transcribe-backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.14
django-cors-headers>=4.3
python-dotenv>=1.0
httpx>=0.27
